+++ src/test/resources/testng_profiles/test_profile_hw4.xml
@@ -1,10 +1,15 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Homework_4-All_Tests" parallel="false">
-    <test name="Homework4">
+<suite name="Homework #4. Allure Report and Jenkins">
+    <test name="Exercise 1, 2 for Allure report Demo">
         <classes>
             <class name="ru.training.at.hw4.tests.ExerciseOne"/>
-<!--            <class name="ru.training.at.hw4.tests.ExerciseTwoTest"/>-->
+            <class name="ru.training.at.hw4.tests.ExerciseTwoTest"/>
+        </classes>
+    </test>
+    <test name="Test that should fail for Allure report Demo">
+        <classes>
+            <class name="ru.training.at.hw4.tests.FailingTest"/>
         </classes>
     </test>
 </suite>
